Move utility functions used only in aesara.scan.basic to aesara.scan.basic
